Fixed heredoc bug.

+++ redirections2.c
@@ -6,7 +6,7 @@
 /*   By: hamad <hamad@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/26 20:05:16 by hamad             #+#    #+#             */
-/*   Updated: 2025/01/07 02:24:22 by hamad            ###   ########.fr       */
+/*   Updated: 2025/01/09 12:00:04 by hamad            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -76,7 +76,11 @@ void	close_files(t_commands *cmds)
 	int	i;
 
 	i = cmds->bfdp;
-	unlink(TEMP_FILE);
+	if (access(TEMP_FILE, F_OK) == 0)
+	{
+		unlink(TEMP_FILE);
+		cmds->hdp = -1;
+	}
 	while (i < cmds->efdp)
 	{
 		close(cmds->fd[i]);
